BUGFIX: Fixed creating HmacHasher failure with unimplemented hash algorithm.

+++ src/ksi/net_http_curl.c
@@ -127,7 +127,7 @@ static size_t receiveDataFromLibCurl(void *ptr, size_t size, size_t nmemb, void 
 	nc->len = bytesCount;
 	tmp_buffer = NULL;
 
-	KSI_LOG_debug(nc->ctx, "0x%x: Received %llu bytes (%llu so far)", nc, (unsigned long long) bytesCount, nc->len);
+	KSI_LOG_debug(nc->ctx, "0x%x: Received %llu bytes (%llu so far)", nc, (unsigned long long) bytesCount, (unsigned long long)nc->len);
 
 	bytesCount = size * nmemb;
